requirements.txt
django
pandas
Faker
djangorestframework
